Add layer variants to combinators.
