+++ tower-util/src/combinator/mod.rs
@@ -4,7 +4,23 @@
 //! with `futures`.
 
 use futures::{future, prelude::*};
+use tower_layer::Layer;
 use tower_service::Service;
+
+mod and_then;
+mod from_err;
+mod inspect;
+mod map;
+mod map_err;
+mod or_else;
+mod then;
+pub use and_then::*;
+pub use from_err::*;
+pub use inspect::*;
+pub use map::*;
+pub use map_err::*;
+pub use or_else::*;
+pub use then::*;
 
 mod sealed {
     pub trait Sealed<T> {}
@@ -135,239 +151,6 @@ where
     }
 }
 
-/// `Service` analog to the `futures::future::Map` combinator.
-pub struct Map<T, F> {
-    inner: T,
-    f: F,
-}
-
-impl<T, F> Map<T, F> {
-    pub const fn new(service: T, f: F) -> Self {
-        Self { inner: service, f }
-    }
-}
-
-impl<Request, T, F, U> Service<Request> for Map<T, F>
-where
-    T: Service<Request>,
-    F: Copy + Fn(T::Response) -> U,
-{
-    type Response = U;
-    type Error = T::Error;
-    type Future = future::Map<T::Future, F>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready()
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        let f = self.f;
-        self.inner.call(req).map(f)
-    }
-}
-
-/// `Service` analog to the `futures::future::MapErr` combinator.
-pub struct MapErr<T, F> {
-    inner: T,
-    f: F,
-}
-
-impl<T, F> MapErr<T, F> {
-    /// Create a new mapped service, wrapping the inner service `service`
-    /// and applying `f` to all errors.
-    pub const fn new(service: T, f: F) -> Self {
-        Self { inner: service, f }
-    }
-}
-
-impl<Request, T, F, U> Service<Request> for MapErr<T, F>
-where
-    T: Service<Request>,
-    F: Copy + Fn(T::Error) -> U,
-{
-    type Response = T::Response;
-    type Error = U;
-    type Future = future::MapErr<T::Future, F>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        let f = self.f;
-        self.inner.poll_ready().map_err(f)
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        let f = self.f;
-        self.inner.call(req).map_err(f)
-    }
-}
-
-/// `Service` analog to the `futures::future::FromErr` combinator.
-pub struct FromErr<T, E> {
-    inner: T,
-    _err_marker: std::marker::PhantomData<E>,
-}
-
-impl<T, E> FromErr<T, E> {
-    pub const fn new(service: T) -> Self {
-        Self {
-            inner: service,
-            _err_marker: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<Request, T, E> Service<Request> for FromErr<T, E>
-where
-    T: Service<Request>,
-    E: From<T::Error>,
-{
-    type Response = T::Response;
-    type Error = E;
-    type Future = future::FromErr<T::Future, E>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready().map_err(std::convert::Into::into)
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        self.inner.call(req).from_err()
-    }
-}
-
-/// `Service` analog to the `futures::future::Then` combinator.
-///
-/// TODO: Currently restricted to not modifying the error type,
-/// since we need `poll_ready` and `call` to have the same type.
-/// But it doesn't make sense to apply `then` to the `poll_ready` type.
-pub struct Then<T, F> {
-    inner: T,
-    f: F,
-}
-
-impl<T, F> Then<T, F> {
-    pub const fn new(service: T, f: F) -> Self {
-        Self { inner: service, f }
-    }
-}
-
-impl<Request, T, F, U> Service<Request> for Then<T, F>
-where
-    T: Service<Request>,
-    F: Copy + FnOnce(Result<T::Response, T::Error>) -> U,
-    U: IntoFuture<Error = T::Error>,
-{
-    type Response = U::Item;
-    type Error = U::Error;
-    type Future = future::Then<T::Future, U, F>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready()
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        let f = self.f;
-        self.inner.call(req).then(f)
-    }
-}
-
-/// `Service` analog to the `futures::future::AndThen` combinator.
-pub struct AndThen<T, F> {
-    inner: T,
-    f: F,
-}
-
-impl<T, F> AndThen<T, F> {
-    pub const fn new(service: T, f: F) -> Self {
-        Self { inner: service, f }
-    }
-}
-
-impl<Request, T, F, U> Service<Request> for AndThen<T, F>
-where
-    T: Service<Request>,
-    F: Copy + FnOnce(T::Response) -> U,
-    U: IntoFuture<Error = T::Error>,
-{
-    type Response = U::Item;
-    type Error = T::Error;
-    type Future = future::AndThen<T::Future, U, F>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready()
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        let f = self.f;
-        self.inner.call(req).and_then(f)
-    }
-}
-
-/// `Service` analog to the `futures::future::OrElse` combinator.
-///
-/// TODO: Currently restricted to not modifying the error type,
-/// since we need `poll_ready` and `call` to have the same type.
-/// But it doesn't make sense to `or_else` the `poll_ready` type.
-pub struct OrElse<T, F> {
-    inner: T,
-    f: F,
-}
-
-impl<T, F> OrElse<T, F> {
-    pub const fn new(service: T, f: F) -> Self {
-        Self { inner: service, f }
-    }
-}
-
-impl<Request, T, F, U> Service<Request> for OrElse<T, F>
-where
-    T: Service<Request>,
-    F: Copy + FnOnce(T::Error) -> U,
-    U: IntoFuture<Item = T::Response, Error = T::Error>,
-{
-    type Response = T::Response;
-    type Error = T::Error;
-    type Future = future::OrElse<T::Future, U, F>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready()
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        let f = self.f;
-        self.inner.call(req).or_else(f)
-    }
-}
-
-/// `Service` analog to the `futures::future::Inspect` combinator.
-pub struct Inspect<T, F> {
-    inner: T,
-    f: F,
-}
-
-impl<T, F> Inspect<T, F> {
-    pub const fn new(service: T, f: F) -> Self {
-        Self { inner: service, f }
-    }
-}
-
-impl<Request, T, F> Service<Request> for Inspect<T, F>
-where
-    T: Service<Request>,
-    F: Copy + FnOnce(&T::Response),
-{
-    type Response = T::Response;
-    type Error = T::Error;
-    type Future = future::Inspect<T::Future, F>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.inner.poll_ready()
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        let f = self.f;
-        self.inner.call(req).inspect(f)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -427,6 +210,17 @@ mod tests {
     fn test_map() {
         let mut mock = tokio_mock_task::MockTask::new();
         let mut svc = Map::new(Foo, |x| format!("Ok-{:?}", x));
+        let mut fut = svc.call(12);
+        let res = assert_ready!(mock.enter(|| fut.poll()));
+        assert_eq!(res, "Ok-12");
+    }
+
+    #[test]
+    fn test_map_layer() {
+        let mut mock = tokio_mock_task::MockTask::new();
+        let mut svc = tower::ServiceBuilder::new()
+            .layer(MapLayer::new(|x| format!("Ok-{:?}", x)))
+            .service(Foo);
         let mut fut = svc.call(12);
         let res = assert_ready!(mock.enter(|| fut.poll()));
         assert_eq!(res, "Ok-12");
